+++ components/wiki/ProfileDetail.module.scss
