+++ components/wiki/ProfileDetails.tsx
