+++ components/wikiList/RecentSearch.module.scss
