+++ components/wikiList/SPinner.tsx
