+++ hooks/wikiList/useRecentSearch.ts
